screen_editor/requirements.txt
pillow==10.4.0
Flask==3.0.3
